requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
redis==5.1.0
rich==13.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
